+++ src/base/func_def_visitor.py
@@ -1,0 +1,13 @@
+from pycparser import c_ast
+
+
+class FuncDefVisitor(c_ast.NodeVisitor):
+    def visit_FuncDef(self, node):
+        print("FuncDef: ", node.decl.name, node.decl.coord)
+        # self.generic_visit(node)
+
+    # def visit_FuncDecl(self, node):
+    #     print("FuncDecl: ", node.args, node.coord)
+
+    # def visit_FuncCall(self, node):
+    #     print("FuncCall: ", node.name.name, node.name.coord)
