+++ src/base/function.py
@@ -1,0 +1,10 @@
+class Function:
+    """関数の値オブジェクト
+    Args:
+        name (str): 関数名
+    """
+
+    def __init__(self, name):
+        self.name = name
+        # self.args = args
+        # self.body = body
